refactor: move database repositories
